Extract profiler state
