+++ src/adapter/11/renderer.ts
@@ -5,6 +5,7 @@ import { PortPageHook } from "../adapter/port";
 import { Renderer } from "../renderer";
 import { Internal } from "./internal";
 import { OptionsV11 } from "./options";
+import { ProfilerState } from "../adapter/profiler";
 
 export interface Preact11Renderer extends Renderer {
 	onCommit(internal: Internal): void;
@@ -16,7 +17,7 @@ export function createV11Renderer(
 	port: PortPageHook,
 	namespace: number,
 	options: OptionsV11,
-	supports: Supports,
+	profiler: ProfilerState,
 	filters: FilterState = DEFAULT_FIlTERS,
 ): Preact11Renderer {
 	return {
